Use 64bit IDs for node retriever

+++ c/usearch.h
@@ -23,7 +23,7 @@ USEARCH_EXPORT typedef uint64_t usearch_key_t;
 // this just sets up an alias so the names are interchangable
 USEARCH_EXPORT typedef usearch_key_t usearch_label_t;
 USEARCH_EXPORT typedef float usearch_distance_t;
-USEARCH_EXPORT typedef void* (*usearch_node_retriever_t)(void* ctx, int index);
+USEARCH_EXPORT typedef void* (*usearch_node_retriever_t)(void* ctx, uint64_t index);
 
 // this is an upper bound. the actual index is smaller.
 #define USEARCH_EMPTY_INDEX_SIZE 512
